wmfdb/cli_admin: Add db_mysql
This is available as db-mysql within tox. It's similar to the mysql.py
from wmfmariadbpy, with the important difference that it takes its own
parameter (the instance in "host:port" format), and options (e.g. to
control logging). All other params+options are forwarded to the mysql
binary. This greatly reduces the complexity of the code.

Bug: T297618
Change-Id: I80134e72734a0c3c9931955065b7174fb1185878