wmfdb/db: Add module for querying databases.
DB is a thin wrapper around pymysql's Connection.
CursorWrapper is a slightly thicker wrapper around pymysql's various
Cursor classes, to provide convenient configuration of query timeouts.

We'll probably want to add a custom QueryResult type to simplify using
the results, but i'm leaving that until i've at least written some code
that uses it as-is.

Bug: T298236
Change-Id: I0b805c2b99866c2fc56d5eef5a90a4daa03cfe26