Configure mysql client preferences for an-redacteddb* hosts
Previously the private wikireplica server used by the Data Engineering
team was called clouddb1021, so the mysql-db utility was able to select
the correct mysql client configuration based on its similar hostname to
the other wikireplica servers.

We have decided to rename this host during a refresh cycle, so its new
name is an-redacteddb1001. This requires a modification of the wmfdb
python library, in order to add this new name as a correct pattern match
for the clientlabsdb options block.

Bug: T368354

See merge request repos/sre/wmfdb!16